iotpython/requirements.txt
cryptography==2.8
flaky==3.6.1
gcp-devrel-py-tools==0.0.15
google-api-python-client==1.8.0
google-auth-httplib2==0.0.3
google-auth==1.14.0
google-cloud-pubsub==1.3.0
google-cloud-iot==1.0.0
grpc-google-iam-v1==0.12.3
pyjwt==1.7.1
paho-mqtt==1.5.0